tts/requirements.txt
sounddevice==0.5.1
soundfile==0.13.1
torch==2.6.0
coqui-tts==0.25.3